[General] Increase max size of folded robot
This change corrects the grammar, cube->box. Because the robot
dimensions are not a symetric cube but a rectangular box.

The height change increases the maximum folded hieght and width to encourage new
teams with small standard research platforms such as the: Toyota HSR and
Fetch.

Max folded hieght: 80cm -> 110cm
Max folded width:  50cm ->  55cm

The unfolded sizes don't need to change as they were large enough.
The rules still state that teams can contact the TC if and request small
exceptions.